When drawing a path, keep it in a separate image.
Then composite it into the previously-drawn paths and background images when the drawing ends. This keeps touch response quick, because it only has to draw one thing.

This does, however, mean that whenever the video recorder wants to capture a frame, we need to composite the two together into a single image before drawing.